change(web): reworks native vs embedded osk definitions
