+++ web/source/kmwembedded.ts
@@ -131,6 +131,24 @@ namespace com.keyman.osk {
     }
   };
 
+  VisualKeyboard.prototype.highlightSubKeys = function(this: VisualKeyboard, k, x, y) {
+    // a dummy function; it's only really used for 'native' KMW.
+  }
+
+  VisualKeyboard.prototype.drawPreview = function(this: VisualKeyboard, w, h, edge) {
+    // a dummy function; it's only really used for 'native' KMW.
+  }
+
+  VisualKeyboard.prototype.addCallout = function(this: VisualKeyboard, key) {
+    // a dummy function; it's only really used for 'native' KMW.
+    return null;
+  }
+
+  VisualKeyboard.prototype.waitForFonts = function(this: VisualKeyboard, kfd, ofd) {
+    // a dummy function; it's only really used for 'native' KMW.
+    return true;
+  }
+
   SuggestionManager.prototype.platformHold = function(this: SuggestionManager, suggestionObj: BannerSuggestion, isCustom: boolean) {
     // Parallels VisualKeyboard.prototype.touchHold, but for predictive suggestions instead of keystrokes.
     let suggestionEle = suggestionObj.div;
